feat(utils): `HashSet` iterator
